% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Create the targets pipeline for the CL-PFU database}
\usage{
get_pipeline(
  input_data_version,
  output_version,
  conn_params,
  countries,
  years,
  do_chops,
  schema_file_path = PFUSetup::get_abs_paths(version =
    input_data_version)[["schema_path"]]
)
}
\arguments{
\item{input_data_version}{A string indicating which input data should be used.}

\item{output_version}{A string indicating the output version.}

\item{conn_params}{A named list containing connection parameters to the database.
See details.}

\item{countries}{A vector of countries to be analyzed.}

\item{years}{A vector of years to be analyzed.}

\item{do_chops}{A boolean that tells whether to perform slices on the
resource matrix and final demand matrix.}
}
\value{
A list representing the \code{targets} pipeline.
}
\description{
This is a type of target factory whose arguments
specify the details of a targets workflow to be constructed
}
\details{
\code{conn_params} should be a named list containing some information for the connection.
\itemize{
\item dbname: a string identifying the name of the database
\item user: the username of the person connecting
\item host: the hostname or IP address of the database management system (DBMS)
\item port: the port for DBMS connections
}

Passwords should \emph{not} be included in \code{conn_params} and will be ignored.
Rather, passwords should be set in the user's \code{.pgpass} file.
}
