% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{pl_destroy}
\alias{pl_destroy}
\title{Reset a database pipeline to original condition}
\usage{
pl_destroy(
  conn = DBI::dbConnect(drv = RPostgres::Postgres(), dbname = conn_args$dbname, host =
    conn_args$host, port = conn_args$port, user = conn_args$user),
  conn_params
)
}
\value{
If successful, \code{TRUE}, otherwise \code{FALSE}.
}
\description{
Sometimes, you just need to start over from scratch.
This function destroys the local \code{targets} cache
and remove all tables in the database at \code{conn}.
This is a very destructive operation, so
be sure you know what you're doing!
}
