% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_schema}
\alias{upload_schema}
\title{Upload a schema for the CL-PFU database}
\usage{
upload_schema(.dm, conn, drop_tables = FALSE)
}
\arguments{
\item{.dm}{A data model from the \code{dm} package to upload.}

\item{conn}{A \code{DBI} connection to a database.}

\item{drop_tables}{A boolean that tells whether to delete
existing tables before uploading the new schema.}
}
\value{
The remote data model
}
\description{
When you need to start over again,
you need to delete database tables and re-upload the schema.
This function makes it easy.
}
\details{
Optionally, deletes existing tables in the database before uploading.

\code{conn}'s user must have superuser privileges.
}
